+++ src/main/java/com/remotelauncher/Constants.java
@@ -8,6 +8,7 @@ package com.remotelauncher;
 
 public class Constants {
 
+    public static final String SERVER_NAME = "localhost";
     public static final int PORT_NUMBER = 81;
 
 }
